Worked on speciation
Generalized a lot of ArrayLists to just Lists
Added the method for adding organisms to the population in such a way as
to group them into species.
Added methods to create new species containing a member of the old
generation as a representative, and fixed the technique for determining
the representative of a species (now does it randomly as it ought to).